Rename Panorama -> Panoramas (#1)
This is the result of

    find . -name '*.py' -exec sed -i 's/Panoramas/Panorama/g' {} \;

on the entire codebase. The reason for doing this is that makemigrations
hasn't been run for years, and running it now will delete the Panorama
table and create a new Panoramas table.

The naming of models in terms of singular/plural is inconsistent anyway,
so this doesn't matter for code style.