Remove the Region class
This class and its table are only relevant for the processing code,
where they are no longer used. The API doesn't look at the regions and
if it needs them, they're still in the object store.

The corresponding migration to remove the database table
panoramas_regions is still to do. makemigrations fails because of an
unmanaged database view.