Database migrations
These drop the Region table and apply various automatically detected
changes. It couldn't use the ManageView class, because it drops and
immediately recreates the adjacencies view, whereas the drop has to be
done first and the recreation last. Every migration going forward will
have to be edited to do the view recreation.

The ManageMaterializedView class was removed because it was unused and
similarly broken.

There are several third-party Django extensions that promise to manage
views, but I could find none that are actively maintained.